refactor: Tidy up code to adhere to Code Quality guidelines
