+++ src/main/java/babe/Main.java
@@ -2,6 +2,7 @@ package babe;
 
 import java.io.IOException;
 
+import babe.controllers.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
