+++ src/main/resources/view/MainWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="babe.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="babe.controllers.MainWindow">
     <children>
         <TextField fx:id="userInput" layoutX="11.0" layoutY="558.0" onAction="#handleUserInput" prefHeight="27.0" prefWidth="306.0" AnchorPane.bottomAnchor="9.0" />
         <Button fx:id="sendButton" layoutX="327.0" layoutY="564.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="27.0" prefWidth="63.0" text="Send" />
